CP-22247: Add temporary vlan bridge to inventory for management vlan.
It is a backport for CA-242016, to fix reset networking performed on
slave for management on VLAN.

Fix will allow xcp-networkd to create temporary vlan bridge in order
to contact master on a vlan network.
On XAPI startup, plugging VLAN PIF will destroy temporary bridge
and create a new vlan bridge.

Signed-off-by: sharad yadav <sharad.yadav@citrix.com>