Use XAPI to edit DNS entries within xsconsole
Instead of directly editing `/etc/resolv.conf`, use XAPI to edit it.
This fixes an issue where the changes weren't retained after a reboot because XAPI
would override the changes with what was stored in its DB. Using XAPI to edit the DNS entries
also applies the changes in XAPI's DB and so changes are kept after reboots.

Signed-off-by: BenjiReis <benjamin.reis@vates.fr>