mypy,pylint: Ignore re-defining class VMUtils in plugins-base/XSFeatureVMCommon.py
mypy and pylint found that VMUtils is already defined by XSConsoleDataUtils.py:
It only provides @staticmethod VMUtils.numLocalResidentVMs(), so it is just a function.

This plugin overwrites the global class for this plugin.
While not great, add a ignore comment to the re-definition for now
with the TODO/FIXME labels added to the comment to fix later.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>