CA-390512: Improved xenapi error handling
In some cases we may fail to query the host CPUs due to an invalid
reference, in this case we should log the failure and not add it into
the list of host CPUs.
In the event of a failure we receive a list of the form:
['HANDLE_INVALID', 'host_cpu', '<uuid>']
In a success we receive a dictionary which is later used to populate the
number of CPUs per name but does not check that the contents of
host_CPUs is a list of dictionaries.