CP-54442: Update for network devices not being renamed
A new proposed change in networkd will replace the interface-rename
functionality to order the network devices. The result of the ordering
performed by interface-rename is to rename the devices like eth0, eth1,
etc. The new change in networkd will not rename the devices any more.
Consequently, the use of hard-coded eth<N> in network reset in xsconsole
should be eliminated to cope with the new change.

Additionally, the new networkd replacing the interface rename
functionality will perform the network reset with xapi together.

The interface rename functionality will be deprecated in the future.

Signed-off-by: Ming Lu <ming.lu@cloud.com>