CA-417641 Update ntp-dhcp server file path
xsconsole can set ntp to Use DHCP NTP Servers or Use Default NTP
Servers, Provide NTP Servers Manually. This is so-called dhcp
default and manual mode.

When switching from other modes to DHCP, the DHCP-provided NTP
servers should be added to the Chrony sources. Conversely, they
should be removed when switching away from DHCP.

In XS9, the ntp-dhcp server path is outdated as the chrony version
update, comparing to XS8.
xs8-chrony: version 3.2, the server file with dhclient is
/var/lib/dhclient/chrony.servers.$interface
xs9-chrony: version 4.6.1, the server file with dhclient is
/run/chrony-dhcp/$interface.sources
So the path should be updated.

Signed-off-by: Changlei Li <changlei.li@cloud.com>